Sets up skeleton work for new command

refactors parse_config_value_set into a more central place
